autolaiskai/autolaiskai/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
crispy-bootstrap5==0.6
cryptography==37.0.4
distro==1.7.0
Django==4.1
django-crispy-forms==1.14.0
docxcompose==1.3.5
docxtpl==0.16.8
h11==0.13.0
idna==3.3
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.23.2
outcome==1.2.0
packaging==23.2
pandas==1.4.4
pdfminer.six==20220524
pdfplumber==0.7.4
phantomjs==1.4.1
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.2.1
regex==2023.12.25
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tabula-py==2.5.1
tabulate==0.8.10
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
Wand==0.6.10
webdriver-manager==4.0.1
wsproto==1.2.0
